feat: party optimizations
Removed many expensive calls to PartyManager#getParty and stored the result in a variable instead.
Removed PartyManager#isInParty calls as it does the same as getParty but returning true/false instead of Party/null, which forces to filter all parties twice. This comes at a small cost of readability, as it makes us use `playerParty != null` instead of `isInParty(player)`.